Adds contents to json templates
Shows the user what the available fields are in a way that is less verbose but more useful than the schemas.

Also allows for the user to copy/paste a template with the appropriate fields already in place, rather than having to refer to the schema constantly to know what is possible, while still having the schema as something to refer to if unsure of anything more.